requirements.txt
# Core web API
fastapi==0.115.0
uvicorn[standard]==0.30.6

# DB + migrations
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
alembic==1.13.2

# Models and validation
pydantic==2.9.2
pydantic-settings==2.4.0

# Task queue / cache
redis==5.0.8
rq==1.16.2
# Or Celery alternative (comment out rq above if you prefer Celery)
# celery==5.4.0

# Agent frameworks (optional for examples/integration)
langchain==0.3.3
openai==1.50.2

# Utilities
python-dotenv==1.0.1
requests==2.32.3

# Security / parsing (optional for rules)
regex==2024.9.11
PyYAML==6.0.2

# Testing
pytest==8.3.3
pytest-cov==5.0.0